Extracting payment success forms to the widgets
Originally invoice and print address forms were hardcoded to the payment
success page including the conditions deciding whether they should be
displayed or not.

Both were extracting to the widgets, keeping the form factories intact
(except the translation strings).

Widgets are directly linked to the payment success page and presenter,
expecting JS and presenter functions to be present to manage AJAX updates
and form validations when directing user out of the page.

remp/crm#664